fix: Use idempotent, collision-resistant algorithm
Previously we used SipHash with a process-specific random key. This was
probably because that is the usual way to use shorthash().

SipHash is suitable for hash tables but we do not have a mechanism for
collisions. Therefore we should be using a collision-resistant hashing
algorithm. libsodium uses Blake2b for the implementation of
generichash().

Also the use of a process-specific random key meant that the hash values
for the same input would differ between different process instances. If
we wish to handle service restarts properly, the hash values must be
consistent over different process instances.